in src:
2009-02-03  Fernando Herrera  <fherrera@novell.com>

	* runtime.cpp:
	* runtime.h: Move moonlight_flags and runtime_flags to 64bits
	as we reached the 32bits limit for debug.

in plugin:


2009-02-03  Fernando Herrera  <fherrera@novell.com>

	* plugin.cpp: moonlight_flags is 64bits now.


svn path=/trunk/moon/; revision=125501

+++ plugin/plugin.cpp
@@ -49,7 +49,7 @@
 // Debug NPStreams
 #define nps(x)//x
 
-extern guint32 moonlight_flags;
+extern guint64 moonlight_flags;
 
 /* gleaned from svn log of the moon module, as well as olive/class/{agclr,agmono,System.Silverlight} */
 static const char *moonlight_authors[] = {
